move and cleanup add_ingots_bulk tests
