+++ tests/main_test.py
@@ -49,135 +49,6 @@ class TestIronForgedBot(unittest.IsolatedAsyncioTestCase):
         self.mock_interaction.guild.members = []
 
     @patch("main.validate_protected_request")
-    def test_addingots_bulk(self, mock_validate_protected_request):
-        """Test that ingots can be added to multiple users."""
-        leader_name = "leader"
-        player1 = "johnnycache"
-        player1_id = 123456
-        player2 = "kennylogs"
-        player2_id = 654321
-
-        mock_validate_protected_request.return_value = (
-            helper_create_member(leader_name, ROLES.LEADERSHIP),
-            leader_name,
-        )
-
-        mock_storage = MagicMock()
-        mock_storage.read_members.return_value = [
-            Member(id=player1_id, runescape_name=player1, ingots=5000),
-            Member(id=player2_id, runescape_name=player2, ingots=400),
-        ]
-
-        mo = mock_open()
-
-        commands = main.IronForgedCommands(MagicMock(), MagicMock(), mock_storage, "")
-
-        with patch("builtins.open", mo):
-            self.loop.run_until_complete(
-                commands.addingotsbulk(
-                    self.mock_interaction, f"{player1},{player2}", 5000
-                )
-            )
-
-        mock_storage.update_members.assert_called_once_with(
-            [
-                Member(id=player1_id, runescape_name=player1, ingots=10000),
-                Member(id=player2_id, runescape_name=player2, ingots=5400),
-            ],
-            leader_name,
-            note="None",
-        )
-
-        mo().write.assert_called_once_with(
-            f"Added 5,000 ingots to {player1}. They now have 10,000 ingots\nAdded 5,000 ingots to {player2}. They now have 5,400 ingots"
-        )
-
-    @patch("main.validate_protected_request")
-    def test_addingots_bulk_whitespace_stripped(self, mock_validate_protected_request):
-        """Test that ingots can be added to multiple users."""
-        leader_name = "leader"
-
-        mock_validate_protected_request.return_value = (
-            helper_create_member(leader_name, ROLES.LEADERSHIP),
-            leader_name,
-        )
-
-        mock_storage = MagicMock()
-        mock_storage.read_members.return_value = [
-            Member(id=123456, runescape_name="johnnycache", ingots=5000),
-            Member(id=654321, runescape_name="kennylogs", ingots=400),
-        ]
-
-        mo = mock_open()
-
-        commands = main.IronForgedCommands(MagicMock(), MagicMock(), mock_storage, "")
-        with patch("builtins.open", mo):
-            # User adds whitespace between args.
-            self.loop.run_until_complete(
-                commands.addingotsbulk(
-                    self.mock_interaction, "johnnycache, skagul tosti", 5000
-                )
-            )
-
-        mock_storage.update_members.assert_called_once_with(
-            [Member(id=123456, runescape_name="johnnycache", ingots=10000)],
-            "leader",
-            note="None",
-        )
-
-        mo().write.assert_called_once_with(
-            """Added 5,000 ingots to johnnycache. They now have 10,000 ingots
-skagul tosti not found in storage."""
-        )
-
-    @patch("main.send_error_response")
-    @patch("main.validate_protected_request")
-    def test_addingots_bulk_player_fail_validation(
-        self, mock_validate_protected_request, mock_send_error_response
-    ):
-        """Test that a missing player is surfaced to caller."""
-        leader_name = "leader"
-        bad_name = "somesuperlongfakename"
-
-        mock_validate_protected_request.return_value = (
-            helper_create_member(leader_name, ROLES.LEADERSHIP),
-            leader_name,
-        )
-
-        mock_storage = MagicMock()
-        mock_storage.read_member.return_value = None
-
-        commands = main.IronForgedCommands(MagicMock(), MagicMock(), mock_storage, "")
-        self.loop.run_until_complete(
-            commands.addingotsbulk(self.mock_interaction, bad_name, 5)
-        )
-
-        mock_send_error_response.assert_awaited_with(
-            self.mock_interaction, "RSN can only be 1-12 characters long"
-        )
-
-    @patch("main.send_error_response")
-    def test_addingots_bulk_permission_denied(self, mock_send_error_response):
-        """Test that non-leadership role can't add ingots."""
-        guild = AsyncMock(discord.Guild)
-        caller = helper_create_member("1eader", ROLES.MEMBER)
-        member = helper_create_member("member", ROLES.MEMBER)
-        guild.members = [caller, member]
-
-        self.mock_interaction.user = caller
-        self.mock_interaction.guild = guild
-
-        commands = main.IronForgedCommands(MagicMock(), MagicMock(), MagicMock(), "")
-        self.loop.run_until_complete(
-            commands.addingotsbulk(self.mock_interaction, caller.name, 500)
-        )
-
-        mock_send_error_response.assert_awaited_with(
-            self.mock_interaction,
-            f"Member '{caller.name}' does not have permission for this action",
-        )
-
-    @patch("main.validate_protected_request")
     def test_updateingots(self, mock_validate_protected_request):
         """Test that ingots can be written for a player."""
         leader_name = "leader"
